refactor(contrib): use @socialgouv/contributions-data
