+++ packages/code-du-travail-data/dataset/fiches_service_public/package.json
@@ -8,6 +8,7 @@
   },
   "dependencies": {
     "@cdt/sources": "^4.14.0",
+    "@socialgouv/contributions-data": "^1.2.0",
     "@socialgouv/fiches-vdd": "^1.24.0",
     "@socialgouv/kali-data": "^1.13.0",
     "@socialgouv/legi-data": "^1.2.0",
