+++ packages/code-du-travail-data/indexing/ breadcrumbs.js
@@ -1,0 +1,19 @@
+import { SOURCES, getRouteBySource } from "@cdt/sources";
+
+export function createThemer(themes) {
+  return function getTheme(slug) {
+    const theme = themes.find((theme) =>
+      theme.refs.some((ref) => ref.url.match(new RegExp(slug)))
+    );
+    let breadcrumbs = [];
+    if (theme) {
+      breadcrumbs = (theme.breadcrumbs || []).concat([
+        {
+          label: theme.title,
+          slug: `/${getRouteBySource(SOURCES.THEMES)}/${theme.slug}`,
+        },
+      ]);
+    }
+    return breadcrumbs;
+  };
+}
