--- packages/code-du-travail-data/lib.js
@@ -1,11 +1,0 @@
-function parseIdcc(idcc) {
-  return parseInt(idcc, 10);
-}
-function formatIdcc(num) {
-  return `0000${num}`.slice(-4);
-}
-
-module.exports = {
-  parseIdcc,
-  formatIdcc,
-};
